refactor: extract message layout policy
